only remove if needed

+++ setup/early/02-gen-answer-file.rb
@@ -50,7 +50,7 @@ q_puppet_enterpriseconsole_inventory_dnsaltnames=MASTER
 
 dashboardhost = nil
 hosts.each do |host|  # Clean up all answer files that might conflict
-  FileUtils.rm ["tmp/answers.#{host}"]
+  FileUtils.rm ["tmp/answers.#{host}"] if File.exists? "tmp/answers.#{host}"
 end
 
 hosts.each do |host|   # find our dashboard host for laster use
